+++ src/app/(usercard)/my-reservation/MyReservationList/MyReservationList.module.scss
@@ -4,3 +4,18 @@
   min-height: 100rem; //케밥이 안보여서 추가
 }
 
+.no-list {
+  margin-top: 10rem;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  gap: 3.69rem;
+  flex: 1;
+}
+
+.no-list-notice {
+@include text(2xl);
+font-weight: 500;
+color: $gray02;
+}
