--- src/app/(usercard)/my-reservation/MyReservationList/NoList.module.scss
@@ -1,15 +1,0 @@
-.no-list {
-  margin-top: 10rem;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  gap: 3.69rem;
-  flex: 1;
-}
-
-.no-list-notice {
-@include text(2xl);
-font-weight: 500;
-color: $gray02;
-}
